Code review: 300730043: Moved event extraction from worker into separate extractor #338
