+++ config/dpkg/changelog
@@ -2,4 +2,4 @@ plaso (1.4.1-1) unstable; urgency=low
 
   * Auto-generated
 
- -- Log2Timeline <log2timeline-dev@googlegroups.com>  Sun, 22 May 2016 19:20:24 +0200
+ -- Log2Timeline <log2timeline-dev@googlegroups.com>  Sun, 22 May 2016 19:53:50 +0200
